+++ index.html
@@ -15,7 +15,7 @@
     <form action="new.php" method="POST">
 
         <div class="field">
-            <input class="text" id="comp" placeholder="New Competition" type="text" name="name"/>
+            <input class="text" id="comp" placeholder="New Competition" type="text" name="name" autocomplete="off"/>
             <p class="helper helper2">Competition Name</p>
         </div>
