+++ utils.php
@@ -23,3 +23,14 @@ function startedCheck(string $id, PDO $pdo): bool
         return false;
     }
 }
+
+function generateRandomString($length = 10): string
+{
+    $characters = '0123456789abcdefghijklmnopqrstuvwxyz';
+    $charactersLength = strlen($characters);
+    $randomString = '';
+    for ($i = 0; $i < $length; $i++) {
+        $randomString .= $characters[random_int(0, $charactersLength - 1)];
+    }
+    return $randomString;
+}
